Unify DocBuilder and BuildQueue
There was no reason to have these be separate. This also gets rid of
some extra open database connections.
