+++ src/lib.rs
@@ -5,7 +5,6 @@
 pub use self::build_queue::BuildQueue;
 pub use self::config::Config;
 pub use self::context::Context;
-pub use self::docbuilder::DocBuilder;
 pub use self::docbuilder::PackageKind;
 pub use self::docbuilder::RustwideBuilder;
 pub use self::index::Index;
